Refactor TaskMapScreen into tasks/map and extract LocationInfoCard component
- Move TaskMapScreen into dedicated ui.datacollection.tasks.map package
- Extract LocationInfo and LocationInfoCard leaf component into tasks.map.components package
- Update AbstractTaskMapFragment with new package imports
- Mirror test structure with TaskMapScreenTest and LocationInfoCardTest in matching packages
