+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/map/TaskMapScreen.kt
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.android.ui.datacollection.tasks
+package org.groundplatform.android.ui.datacollection.tasks.map
 
-import androidx.annotation.StringRes
 import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
@@ -25,21 +24,12 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.groundplatform.android.R
@@ -47,29 +37,18 @@ import org.groundplatform.android.ui.common.ExcludeFromJacocoGeneratedReport
 import org.groundplatform.android.ui.components.MapFloatingActionButton
 import org.groundplatform.android.ui.components.MapFloatingActionButtonType
 import org.groundplatform.android.ui.components.RecenterButton
+import org.groundplatform.android.ui.datacollection.tasks.map.components.LocationInfo
+import org.groundplatform.android.ui.datacollection.tasks.map.components.LocationInfoCard
 import org.groundplatform.ui.theme.AppTheme
 
 @VisibleForTesting
 object TaskMapScreenTestTags {
   const val CENTER_MARKER = "task_map_center_marker"
-  const val LOCATION_INFO_CARD = "task_map_location_info_card"
-  const val CURRENT_LOCATION_TITLE = "task_map_current_location_title"
-  const val CURRENT_LOCATION_VALUE = "task_map_current_location_value"
-  const val ACCURACY_TITLE = "task_map_accuracy_title"
-  const val ACCURACY_VALUE = "task_map_accuracy_value"
 }
 
-@Immutable
-data class LocationInfo(
-  @StringRes val titleRes: Int,
-  val locationText: String,
-  val accuracyText: String? = null,
-  val isAccuracyGood: Boolean = false,
-)
-
 /**
- * Screen overlay for map-based tasks, including center crosshair, map type button, recenter button,
- * location lock button, and location info card.
+ * Screen overlay for map-based tasks, including center crosshairs, map type button, recenter
+ * button, location lock button, and location info card.
  */
 @Composable
 fun TaskMapScreen(
@@ -119,66 +98,6 @@ fun TaskMapScreen(
         }
 
         MapFloatingActionButton(type = locationLockButtonType, onClick = onLocationLockClicked)
-      }
-    }
-  }
-}
-
-@Composable
-fun LocationInfoCard(locationInfo: LocationInfo, modifier: Modifier = Modifier) {
-  Card(
-    modifier = modifier.testTag(TaskMapScreenTestTags.LOCATION_INFO_CARD),
-    shape = RoundedCornerShape(8.dp),
-    colors = CardDefaults.cardColors(containerColor = Color.White),
-  ) {
-    Row(
-      modifier = Modifier.fillMaxWidth().padding(8.dp),
-      verticalAlignment = Alignment.CenterVertically,
-    ) {
-      Column(
-        modifier =
-          if (locationInfo.accuracyText != null) {
-            Modifier.weight(0.75f)
-          } else {
-            Modifier.fillMaxWidth()
-          }
-      ) {
-        Text(
-          text = stringResource(locationInfo.titleRes),
-          style = MaterialTheme.typography.labelSmall,
-          color = Color(0xFF5E5E5E),
-          modifier = Modifier.testTag(TaskMapScreenTestTags.CURRENT_LOCATION_TITLE),
-        )
-        Text(
-          text = locationInfo.locationText,
-          style = MaterialTheme.typography.labelMedium,
-          color = Color(0xFF424940),
-          modifier = Modifier.testTag(TaskMapScreenTestTags.CURRENT_LOCATION_VALUE),
-        )
-      }
-
-      if (locationInfo.accuracyText != null) {
-        Column(modifier = Modifier.weight(0.25f)) {
-          Text(
-            text = stringResource(R.string.accuracy),
-            style = MaterialTheme.typography.labelSmall,
-            color = Color(0xFF5E5E5E),
-            modifier = Modifier.testTag(TaskMapScreenTestTags.ACCURACY_TITLE),
-          )
-          Text(
-            text = locationInfo.accuracyText,
-            style = MaterialTheme.typography.labelMedium,
-            color =
-              colorResource(
-                if (locationInfo.isAccuracyGood) {
-                  R.color.accuracy_good
-                } else {
-                  R.color.accuracy_bad
-                }
-              ),
-            modifier = Modifier.testTag(TaskMapScreenTestTags.ACCURACY_VALUE),
-          )
-        }
       }
     }
   }
